+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Redirect Changelog
-### Unreleased
 
+### 2.0.12 - 2019-07-12
 ### Added
 - We now keep track of the last referrer for a 404